Update imports for moved elements.
