+++ app/src/main/java/com/hfad/palamarchuksuperapp/data/repository/MessageGroupRepositoryImpl.kt
@@ -1,10 +1,11 @@
 package com.hfad.palamarchuksuperapp.data.repository
 
+import com.hfad.palamarchuksuperapp.core.data.withSqlErrorHandling
 import com.hfad.palamarchuksuperapp.data.dao.MessageGroupDao
 import com.hfad.palamarchuksuperapp.data.entities.MessageGroupEntity
-import com.hfad.palamarchuksuperapp.domain.models.AppError
+import com.hfad.palamarchuksuperapp.core.domain.AppError
 import com.hfad.palamarchuksuperapp.domain.models.MessageGroup
-import com.hfad.palamarchuksuperapp.domain.models.Result
+import com.hfad.palamarchuksuperapp.core.domain.Result
 import com.hfad.palamarchuksuperapp.domain.repository.MessageGroupRepository
 import javax.inject.Inject
