--- app/src/main/java/com/hfad/palamarchuksuperapp/domain/models/Error.kt
@@ -1,3 +1,0 @@
-package com.hfad.palamarchuksuperapp.domain.models
-
-sealed interface Error
