+++ app/src/main/java/com/hfad/palamarchuksuperapp/domain/repository/AiModelHandler.kt
@@ -4,8 +4,8 @@ import com.hfad.palamarchuksuperapp.domain.models.AiModel
 import com.hfad.palamarchuksuperapp.domain.models.MessageGroup
 import com.hfad.palamarchuksuperapp.domain.models.MessageAI
 import com.hfad.palamarchuksuperapp.domain.models.AiHandlerInfo
-import com.hfad.palamarchuksuperapp.domain.models.AppError
-import com.hfad.palamarchuksuperapp.domain.models.Result
+import com.hfad.palamarchuksuperapp.core.domain.AppError
+import com.hfad.palamarchuksuperapp.core.domain.Result
 import kotlinx.collections.immutable.PersistentList
 import kotlinx.coroutines.flow.StateFlow
