+++ app/src/main/java/com/hfad/palamarchuksuperapp/domain/repository/MessageAiRepository.kt
@@ -2,9 +2,9 @@ package com.hfad.palamarchuksuperapp.domain.repository
 
 import com.hfad.palamarchuksuperapp.data.entities.MessageAiEntity
 import com.hfad.palamarchuksuperapp.data.entities.MessageStatus
-import com.hfad.palamarchuksuperapp.domain.models.AppError
+import com.hfad.palamarchuksuperapp.core.domain.AppError
 import com.hfad.palamarchuksuperapp.domain.models.MessageAI
-import com.hfad.palamarchuksuperapp.domain.models.Result
+import com.hfad.palamarchuksuperapp.core.domain.Result
 
 interface MessageAiRepository {
     suspend fun addMessageAiEntity(messageAiEntity: MessageAiEntity): Result<Long, AppError>
