+++ app/src/main/java/com/hfad/palamarchuksuperapp/domain/repository/MessageGroupRepository.kt
@@ -1,9 +1,9 @@
 package com.hfad.palamarchuksuperapp.domain.repository
 
 import com.hfad.palamarchuksuperapp.data.entities.MessageGroupEntity
-import com.hfad.palamarchuksuperapp.domain.models.AppError
+import com.hfad.palamarchuksuperapp.core.domain.AppError
 import com.hfad.palamarchuksuperapp.domain.models.MessageGroup
-import com.hfad.palamarchuksuperapp.domain.models.Result
+import com.hfad.palamarchuksuperapp.core.domain.Result
 
 interface MessageGroupRepository {
     suspend fun getMessageGroup(chatId: Int): Result<MessageGroup, AppError>
