+++ app/src/main/java/com/hfad/palamarchuksuperapp/domain/usecases/ChooseMessageAiUseCase.kt
@@ -1,8 +1,8 @@
 package com.hfad.palamarchuksuperapp.domain.usecases
 
-import com.hfad.palamarchuksuperapp.domain.models.AppError
+import com.hfad.palamarchuksuperapp.core.domain.AppError
 import com.hfad.palamarchuksuperapp.domain.models.MessageAI
-import com.hfad.palamarchuksuperapp.domain.models.Result
+import com.hfad.palamarchuksuperapp.core.domain.Result
 import com.hfad.palamarchuksuperapp.domain.repository.ChatController
 import javax.inject.Inject
