+++ app/src/main/java/com/hfad/palamarchuksuperapp/domain/usecases/SendChatRequestUseCase.kt
@@ -1,11 +1,11 @@
 package com.hfad.palamarchuksuperapp.domain.usecases
 
 import com.hfad.palamarchuksuperapp.data.entities.MessageStatus
-import com.hfad.palamarchuksuperapp.domain.models.AppError
+import com.hfad.palamarchuksuperapp.core.domain.AppError
 import com.hfad.palamarchuksuperapp.domain.models.MessageAI
 import com.hfad.palamarchuksuperapp.domain.models.MessageGroup
 import com.hfad.palamarchuksuperapp.domain.models.MessageType
-import com.hfad.palamarchuksuperapp.domain.models.Result
+import com.hfad.palamarchuksuperapp.core.domain.Result
 import com.hfad.palamarchuksuperapp.domain.models.Role
 import com.hfad.palamarchuksuperapp.domain.repository.AiModelHandler
 import com.hfad.palamarchuksuperapp.domain.repository.ChatController
