+++ app/src/main/java/com/hfad/palamarchuksuperapp/ui/compose/AiHandlerScreen.kt
@@ -59,14 +59,14 @@ import com.hfad.palamarchuksuperapp.domain.models.AiHandlerInfo
 import com.hfad.palamarchuksuperapp.domain.models.AiModel
 import com.hfad.palamarchuksuperapp.domain.models.LLMName
 import com.hfad.palamarchuksuperapp.domain.repository.AiModelHandler
-import com.hfad.palamarchuksuperapp.ui.reusable.AppDialog
-import com.hfad.palamarchuksuperapp.ui.reusable.ConfirmationDialog
-import com.hfad.palamarchuksuperapp.ui.reusable.elements.AppOutlinedTextDialogField
-import com.hfad.palamarchuksuperapp.ui.reusable.elements.AppOutlinedTextField
-import com.hfad.palamarchuksuperapp.ui.reusable.elements.AppOutlinedTextPopUpField
-import com.hfad.palamarchuksuperapp.ui.reusable.elements.AppText
-import com.hfad.palamarchuksuperapp.ui.reusable.elements.appEditOutlinedTextConfig
-import com.hfad.palamarchuksuperapp.ui.reusable.elements.appTextConfig
+import com.hfad.palamarchuksuperapp.core.ui.composables.elements.AppDialog
+import com.hfad.palamarchuksuperapp.core.ui.composables.elements.ConfirmationDialog
+import com.hfad.palamarchuksuperapp.core.ui.composables.basic.AppOutlinedTextDialogField
+import com.hfad.palamarchuksuperapp.core.ui.composables.basic.AppOutlinedTextField
+import com.hfad.palamarchuksuperapp.core.ui.composables.basic.AppOutlinedTextPopUpField
+import com.hfad.palamarchuksuperapp.core.ui.composables.basic.AppText
+import com.hfad.palamarchuksuperapp.core.ui.composables.basic.appEditOutlinedTextConfig
+import com.hfad.palamarchuksuperapp.core.ui.composables.basic.appTextConfig
 import com.hfad.palamarchuksuperapp.ui.viewModels.ChatBotViewModel
 import io.ktor.client.HttpClient
 import kotlinx.collections.immutable.PersistentList
