+++ core/src/main/java/com/hfad/palamarchuksuperapp/core/domain/Error.kt
@@ -1,0 +1,3 @@
+package com.hfad.palamarchuksuperapp.core.domain
+
+sealed interface Error
