+++ core/src/main/java/com/hfad/palamarchuksuperapp/core/ui/genericViewModel/GenericViewModel.kt
@@ -3,7 +3,7 @@ package com.hfad.palamarchuksuperapp.core.ui.genericViewModel
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.hfad.palamarchuksuperapp.domain.models.AppError
+import com.hfad.palamarchuksuperapp.core.domain.AppError
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
